requirements.txt
streamlit
pandas==2.1.0
requests
matplotlib==3.8.0
python-dotenv==1.1.1
pandas
matplotlib